Move convolutional conditioners
