+++ torchflows/bijections/finite/multiscale/conditioning/classic.py
@@ -1,8 +1,9 @@
-import math
 from typing import Tuple
 
 import torch
 import torch.nn as nn
+
+from torchflows.bijections.finite.autoregressive.conditioning.transforms import TensorConditionerTransform
 
 
 class ConvModifier(nn.Module):
@@ -134,10 +135,25 @@ class ConvNet(nn.Module):
         return x
 
 
-if __name__ == '__main__':
-    torch.manual_seed(0)
-    im_shape = (1, 36, 29)
-    images = torch.randn(size=(11, *im_shape))
-    net = ConvNet(input_shape=im_shape, n_outputs=77)
-    out = net(images)
-    print(f'{out.shape = }')
+class ConvNetConditioner(TensorConditionerTransform):
+    def __init__(self,
+                 input_event_shape: torch.Size,
+                 parameter_shape: torch.Size,
+                 kernels: Tuple[int, ...] = None,
+                 **kwargs):
+        super().__init__(
+            input_event_shape=input_event_shape,
+            context_shape=None,
+            parameter_shape=parameter_shape,
+            **kwargs
+        )
+        self.network = ConvNet(
+            input_shape=input_event_shape,
+            n_outputs=self.n_transformer_parameters,
+            kernels=kernels,
+            # output_lower_bound=-8.0,
+            # output_upper_bound=8.0,
+        )
+
+    def predict_theta_flat(self, x: torch.Tensor, context: torch.Tensor = None) -> torch.Tensor:
+        return self.network(x)
